refactor(scheduler): convert opt-in crate into a datafusion-examples example
Address PR #23282 review feedback: fold the datafusion-scheduler crate into
datafusion-examples/examples/scheduler/ as a runnable example, with all
integration tests preserved as `#[cfg(test)]` modules so `cargo test --example
scheduler` still exercises the isolated / serialized execution path.

- New example at `datafusion-examples/examples/scheduler/` mirrors the
  original crate layout (config, exchange/{sink,source,codec}, executor,
  scheduler, serde, stage, test_util) plus a `main.rs` runner exposing a
  `distributed_pipeline` subcommand.
- All tests folded inline: stage-splitting tests in `scheduler.rs`, exchange
  round-trip / backpressure / codec tests in `exchange/{mod,codec}.rs`, and
  the equivalence / error / single-stage / two-stage / streaming suites as
  separate test modules in `main.rs`. Same 18 passing + 2 ignored tests as
  before, plus a smoke test for the runnable example.
- Remove the `datafusion-scheduler/` crate and its workspace `exclude`.
- Add a "Scheduler Examples" entry to `datafusion-examples/README.md`.
