--- datafusion-scheduler/Cargo.toml
@@ -1,40 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-[package]
-name = "datafusion-scheduler"
-description = "In-process model of stage-based distributed execution for Apache DataFusion (test/regression harness)"
-version = "0.1.0"
-edition = "2024"
-rust-version = "1.88.0"
-license = "Apache-2.0"
-publish = false
-
-[dependencies]
-datafusion = { path = "../datafusion/core" }
-datafusion-common = { path = "../datafusion/common" }
-datafusion-common-runtime = { path = "../datafusion/common-runtime" }
-datafusion-proto = { path = "../datafusion/proto" }
-futures = "0.3"
-tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
-prost = "0.14.1"                                                      # pinned to match datafusion-proto's resolved prost (verified via `cargo tree -i prost`)
-
-[dev-dependencies]
-tempfile = "3" # only tests need a temp dir now (the #1907 CSV-scan regression guard)
-
-[lints.clippy]
-all = "warn"
